feat: prepare for meger
